Only prefix expression when it's a single char
Instead of only looking at the first character in the expression, only
prefix the answer if it's the only character in the expression. This was
causing a problem if the answer was negative and the user typed a minus
sign.

It will still add the previous answer if you type a minus sign, but you
can now at least enter what you wanted, then delete the answer prefix.